% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_baing.R
\name{f_baing}
\alias{f_baing}
\title{f_baing}
\usage{
f_baing(x, Nr_max, ic)
}
\arguments{
\item{x}{Nt x Nn matrix}

\item{Nr_max}{maximum number of factors}

\item{ic}{information criterion, if "none", Nr = Nr_max!}
}
\value{
Number of factors selected by the chosen criterion
}
\description{
Selects the number of factors according to the information criteria 
described in Bai and Ng (2002, Econometrica)
}
\examples{
Nr <- f_baing(x, Nr_max = 10, ic = "PC_p1")

}
