% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_emalg.R
\name{f_emalg}
\alias{f_emalg}
\title{f_emalg}
\usage{
f_emalg(x, Nr_max = 10, Niter = 50, ic = "PC_p1",
  print_iter = TRUE)
}
\arguments{
\item{x}{Nt x Nn matrix. NA are allowed}

\item{Nr_max}{maximum number of factors. Defaults to 10.}

\item{Niter}{maximum number of iterations. Defaults to 50.}

\item{ic}{information criterion. Defaults to "PC_p1".}

\item{print_iter}{If TRUE, display the current iteration, the value of the error and the number of chosen factors. Defaults to TRUE.}
}
\value{
Tidy dataframe
}
\description{
Estimates the common factors of x following an Expectation-Maximization algorithm
as described in e.g. Stock and Watson (2002, JBES)#'
}
\examples{
fredmd <- f_loadfredmd(vintage = "2019-09", keep_vars = c("INDPRO", "PAYEMS", "CPIAUCSL"))

}
