% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_loadfredmd.R
\name{f_loadfredmd}
\alias{f_loadfredmd}
\title{f_loadfredmd}
\usage{
f_loadfredmd(vintage, dirname = NULL, alternative_trafos = NULL,
  sample_start = "1959-01-01", sample_end = NULL, keep_vars = NULL,
  remove_vars = NULL, drop_firstobs = TRUE)
}
\arguments{
\item{vintage}{String of the form "yyyy-mm"}

\item{dirname}{location of csv-file. Defaults to NULL.}

\item{alternative_trafos}{Vector containing values with which to overwrite the original transformation codes. Defaults to NULL}

\item{sample_start}{String of the form "yyyy-mm-dd" containing the start date of the sample. Defaults to "1959-01-01".}

\item{keep_vars}{Vector of strings containing variables to keep, all others are dropped. Defaults to NULL.}

\item{remove_vars}{Vector of strings containing variables to remove, all others are kept. Defaults to NULL.}

\item{drop_firstobs}{Remove first two observations that contain many NA's due to transformations. Defaults to TRUE.}

\item{sample_start}{String of the form "yyyy-mm-dd" containing the end date of the sample. Defaults to NULL, i.e. all available observations.}
}
\value{
Tidy dataframe
}
\description{
Loads, transforms and tidies up the FRED-MD data.
}
\examples{
fredmd <- f_loadfredmd(vintage = "2019-09", keep_vars = c("INDPRO", "PAYEMS", "CPIAUCSL"))

}
