% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_pca.R
\name{f_pca}
\alias{f_pca}
\title{f_pca}
\usage{
f_pca(x, Nr)
}
\arguments{
\item{x}{Nt x Nn matrix.}
}
\value{
List containing the first Nr principle components (f), 
the associated loadings (lam), the common and idiosyncratic 
components (chi and e, respectively) as well as the 
eigenvalues (eigvals) of the covariance matrix of x
}
\description{
Performs principle components analysis on x.
}
\examples{
tmp <- f_pca(x, Nr = 8)
f <- tmp$f # factors
lam <- tmp$lam # loadings

}
