% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_removeoutl.R
\name{f_removeoutl}
\alias{f_removeoutl}
\title{f_removeoutl}
\usage{
f_removeoutl(x)
}
\arguments{
\item{x}{Nt x Nn matrix}
}
\value{
List containing the data matrix in which outliers have been set to NA (x_outl) 
as well as a count of the number of outliers per series.
}
\description{
Removes outliers from input matrix by setting them to NA. Outliers are 
defined as any observation that is larger than 10 times the 
interquartile range (in absolute value).
}
\examples{
tmp <- f_removeoutl(x)
x_outl <- tmp$x_outl
n_outl <- tmp$n_outl

}
